+++ examples/refsi/refsi_g1_wi/compiler/refsi_g1_wi/source/refsi_pass_machinery.cpp
@@ -29,6 +29,7 @@
 #include <compiler/utils/simple_callback_pass.h>
 #include <llvm/ADT/StringRef.h>
 #include <llvm/Passes/PassBuilder.h>
+#include <llvm/Transforms/Utils/Cloning.h>
 #include <metadata/handler/vectorize_info_metadata.h>
 #include <multi_llvm/optional_helper.h>
 #include <refsi_g1_wi/refsi_pass_machinery.h>