+++ modules/compiler/source/base/source/base_module_pass_machinery.cpp
@@ -41,6 +41,7 @@
 #include <compiler/utils/fixup_calling_convention_pass.h>
 #include <compiler/utils/handle_barriers_pass.h>
 #include <compiler/utils/link_builtins_pass.h>
+#include <compiler/utils/lower_to_mux_builtins_pass.h>
 #include <compiler/utils/make_function_name_unique_pass.h>
 #include <compiler/utils/materialize_absent_work_item_builtins_pass.h>
 #include <compiler/utils/metadata_analysis.h>
@@ -55,9 +56,7 @@
 #include <compiler/utils/replace_address_space_qualifier_functions_pass.h>
 #include <compiler/utils/replace_async_copies_pass.h>
 #include <compiler/utils/replace_atomic_funcs_pass.h>
-#include <compiler/utils/replace_barriers_pass.h>
 #include <compiler/utils/replace_c11_atomic_funcs_pass.h>
-#include <compiler/utils/replace_group_funcs_pass.h>
 #include <compiler/utils/replace_local_module_scope_variables_pass.h>
 #include <compiler/utils/replace_mem_intrinsics_pass.h>
 #include <compiler/utils/replace_mux_math_decls_pass.h>
@@ -142,9 +141,9 @@ Expected<StringRef> parseMakeFunctionNameUniquePassOptions(StringRef Params) {
 }
 
 template <size_t N>
-static ErrorOr<std::array<multi_llvm::Optional<uint64_t>, N>> parseIntList(
+static ErrorOr<std::array<std::optional<uint64_t>, N>> parseIntList(
     StringRef OptionVal, bool AllowNegative = false) {
-  std::array<multi_llvm::Optional<uint64_t>, N> Arr;
+  std::array<std::optional<uint64_t>, N> Arr;
   for (unsigned i = 0; i < N; i++) {
     int64_t Res;
     StringRef Val;
@@ -165,7 +164,7 @@ static ErrorOr<std::array<multi_llvm::Optional<uint64_t>, N>> parseIntList(
   if (!OptionVal.empty()) {
     return std::errc::argument_list_too_long;
   }
-  return std::array<multi_llvm::Optional<uint64_t>, 3U>(Arr);
+  return std::array<std::optional<uint64_t>, 3U>(Arr);
 }
 
 constexpr const char LocalSizesOptName[] = "max-local-sizes=";
@@ -179,7 +178,7 @@ parseEncodeBuiltinRangeMetadataPassOptions(StringRef Params) {
     std::tie(ParamName, Params) = Params.split(';');
 
     StringRef OptName;
-    std::array<multi_llvm::Optional<uint64_t>, 3> *SizesPtr = nullptr;
+    std::array<std::optional<uint64_t>, 3> *SizesPtr = nullptr;
     if (ParamName.consume_front(LocalSizesOptName)) {
       OptName = LocalSizesOptName;
       SizesPtr = &Opts.MaxLocalSizes;